server/requirements.txt
faster-whisper>=0.10.0
numpy>=1.24.0
PyYAML>=6.0
transformers>=4.35.0
torch>=2.0.0
accelerate>=0.24.0
edge-tts>=6.1.0
librosa>=0.10.0
soundfile>=0.12.0
requests>=2.31.0
feedparser>=6.0.10
python-dotenv>=1.0.0
pytest>=7.0.0